# contracts/generated/openapi.json
{
  "components": {
    "schemas": {
      "access_v1_ActivateTenantMemberRequest": {
        "properties": {
          "userId": {
            "type": "string"
          },
          "version": {
            "format": "uint64",
            "type": "string",
            "x-yunka-protobuf-json-integer": true
          }
        },
        "type": "object",
        "x-yunka-proto-full-name": "access.v1.ActivateTenantMemberRequest"
      },
...
      "access_v1_InviteTenantMemberRequest": {
        "properties": {
          "email": {
            "type": "string"
          }
        },
        "type": "object",
        "x-yunka-proto-full-name": "access.v1.InviteTenantMemberRequest"
      },
...
      "access_v1_ListTenantMembersResponse": {
        "properties": {
          "members": {
            "items": {
              "$ref": "#/components/schemas/access_v1_TenantMemberDTO"
            },
            "type": "array"
          }
        },
        "type": "object",
        "x-yunka-proto-full-name": "access.v1.ListTenantMembersResponse"
      },
...
      "access_v1_RemoveTenantMemberRequest": {
        "properties": {
          "userId": {
            "type": "string"
          },
          "version": {
            "format": "uint64",
            "type": "string",
            "x-yunka-protobuf-json-integer": true
          }
        },
        "type": "object",
        "x-yunka-proto-full-name": "access.v1.RemoveTenantMemberRequest"
      },
...
      "access_v1_SuspendTenantMemberRequest": {
        "properties": {
          "userId": {
            "type": "string"
          },
          "version": {
            "format": "uint64",
            "type": "string",
            "x-yunka-protobuf-json-integer": true
          }
        },
        "type": "object",
        "x-yunka-proto-full-name": "access.v1.SuspendTenantMemberRequest"
      },
...
      "access_v1_TenantMemberDTO": {
        "properties": {
          "email": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/access_v1_TenantMemberStatus"
          },
          "userId": {
            "type": "string"
          },
          "version": {
            "format": "uint64",
            "type": "string",
            "x-yunka-protobuf-json-integer": true
          }
        },
        "type": "object",
        "x-yunka-proto-full-name": "access.v1.TenantMemberDTO"
      },
      "access_v1_TenantMemberStatus": {
        "enum": [
          "TENANT_MEMBER_STATUS_UNSPECIFIED",
          "TENANT_MEMBER_STATUS_INVITED",
          "TENANT_MEMBER_STATUS_ACTIVE",
          "TENANT_MEMBER_STATUS_SUSPENDED",
          "TENANT_MEMBER_STATUS_REMOVED"
        ],
        "type": "string",
        "x-yunka-proto-full-name": "access.v1.TenantMemberStatus"
      },
...
    }
  },
  "info": {
    "title": "yunka API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/tenant/members": {
      "get": {
        "operationId": "access.v1.TenantMemberLifecycleApplication.ListTenantMembers",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/access_v1_ListTenantMembersResponse"
                }
              }
            },
            "description": "Successful response"
          }
        },
        "tags": [
          "access.v1.TenantMemberLifecycleApplication"
        ],
        "x-yunka-rpc": "/access.v1.TenantMemberLifecycleApplication/ListTenantMembers"
      },
      "post": {
        "operationId": "access.v1.TenantMemberLifecycleApplication.InviteTenantMember",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/access_v1_InviteTenantMemberRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/access_v1_TenantMemberDTO"
                }
              }
            },
            "description": "Successful response"
          }
        },
        "tags": [
          "access.v1.TenantMemberLifecycleApplication"
        ],
        "x-yunka-rpc": "/access.v1.TenantMemberLifecycleApplication/InviteTenantMember"
      }
    },
    "/v1/tenant/members/{user_id}": {
      "get": {
        "operationId": "access.v1.TenantMemberLifecycleApplication.GetTenantMember",
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/access_v1_TenantMemberDTO"
                }
              }
            },
            "description": "Successful response"
          }
        },
        "tags": [
          "access.v1.TenantMemberLifecycleApplication"
        ],
        "x-yunka-rpc": "/access.v1.TenantMemberLifecycleApplication/GetTenantMember"
      }
    },
    "/v1/tenant/members/{user_id}/activate": {
      "post": {
        "operationId": "access.v1.TenantMemberLifecycleApplication.ActivateTenantMember",
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/access_v1_ActivateTenantMemberRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/access_v1_TenantMemberDTO"
                }
              }
            },
            "description": "Successful response"
          }
        },
        "tags": [
          "access.v1.TenantMemberLifecycleApplication"
        ],
        "x-yunka-rpc": "/access.v1.TenantMemberLifecycleApplication/ActivateTenantMember"
      }
    },
    "/v1/tenant/members/{user_id}/remove": {
      "post": {
        "operationId": "access.v1.TenantMemberLifecycleApplication.RemoveTenantMember",
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/access_v1_RemoveTenantMemberRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/access_v1_TenantMemberDTO"
                }
              }
            },
            "description": "Successful response"
          }
        },
        "tags": [
          "access.v1.TenantMemberLifecycleApplication"
        ],
        "x-yunka-rpc": "/access.v1.TenantMemberLifecycleApplication/RemoveTenantMember"
      }
    },
    "/v1/tenant/members/{user_id}/suspend": {
      "post": {
        "operationId": "access.v1.TenantMemberLifecycleApplication.SuspendTenantMember",
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/access_v1_SuspendTenantMemberRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/access_v1_TenantMemberDTO"
                }
              }
            },
            "description": "Successful response"
          }
        },
        "tags": [
          "access.v1.TenantMemberLifecycleApplication"
        ],
        "x-yunka-rpc": "/access.v1.TenantMemberLifecycleApplication/SuspendTenantMember"
      }
    },
...
  },
...
}