# internal/access/transport/rest/zz_yunka_tenant_member_lifecycle_operation_executor_gen.go
// Code generated by yunka contract; DO NOT EDIT.

package rest

import (
	"errors"
	accessv1 "github.com/hvritual/biz/contracts/gen/access/v1"
	application "github.com/hvritual/biz/internal/access/application"
	policy "github.com/hvritual/biz/internal/access/policy"
	codes "google.golang.org/grpc/codes"
	status "google.golang.org/grpc/status"
	protojson "google.golang.org/protobuf/encoding/protojson"
	io "io"
	"net/http"
	execution "yunka.io/framework/execution"
	operation "yunka.io/framework/operation"
	authz "yunka.io/gateway/authz"
	httpbinding "yunka.io/gateway/httpbinding"
)

type TenantMemberLifecycleOperationHandler struct {
	application application.TenantMemberLifecycleApplication
	executor    operation.Executor
}

func RegisterTenantMemberLifecycleOperationExecutor(mux *http.ServeMux, application application.TenantMemberLifecycleApplication, executor operation.Executor) error {
	if mux == nil {
		return errors.New("contract C9 REST adapter: mux is required")
	}
	if application == nil {
		return errors.New("contract C9 REST adapter: application is required")
	}
	if executor == nil {
		return errors.New("contract C9 REST adapter: operation executor is required")
	}
	handler := &TenantMemberLifecycleOperationHandler{application: application, executor: executor}
	if err := httpbinding.Register(mux, "POST", "/v1/tenant/members/{user_id}/activate", handler.handleOperationActivateTenantMember); err != nil {
		return err
	}
	if err := httpbinding.Register(mux, "GET", "/v1/tenant/members/{user_id}", handler.handleOperationGetTenantMember); err != nil {
		return err
	}
	if err := httpbinding.Register(mux, "POST", "/v1/tenant/members", handler.handleOperationInviteTenantMember); err != nil {
		return err
	}
	if err := httpbinding.Register(mux, "GET", "/v1/tenant/members", handler.handleOperationListTenantMembers); err != nil {
		return err
	}
	if err := httpbinding.Register(mux, "POST", "/v1/tenant/members/{user_id}/remove", handler.handleOperationRemoveTenantMember); err != nil {
		return err
	}
	if err := httpbinding.Register(mux, "POST", "/v1/tenant/members/{user_id}/suspend", handler.handleOperationSuspendTenantMember); err != nil {
		return err
	}
	return nil
}

func writeTenantMemberLifecycleOperationError(writer http.ResponseWriter, err error) {
	if authz.IsDenied(err) {
		statusCode := http.StatusForbidden
		var denied *authz.DeniedError
		if errors.As(err, &denied) && (denied.Decision.Reason == authz.ReasonUnauthenticated || denied.Decision.Reason == authz.ReasonAuthenticationMethod) {
			statusCode = http.StatusUnauthorized
		}
		http.Error(writer, http.StatusText(statusCode), statusCode)
		return
	}
	if errors.Is(err, execution.ErrIdempotencyKeyRequired) {
		http.Error(writer, "idempotency key required", http.StatusBadRequest)
		return
	}
	if errors.Is(err, execution.ErrIdempotencyInProgress) || errors.Is(err, execution.ErrIdempotencyCompleted) {
		http.Error(writer, "idempotency conflict", http.StatusConflict)
		return
	}
	if code := status.Code(err); code == codes.Aborted || code == codes.AlreadyExists {
		http.Error(writer, "application conflict", http.StatusConflict)
		return
	}
	if errors.Is(err, operation.ErrExecutorUnavailable) || errors.Is(err, operation.ErrSecurityUnavailable) || errors.Is(err, operation.ErrSecurityNilContext) || errors.Is(err, operation.ErrIdempotencyUnavailable) {
		http.Error(writer, "operation execution unavailable", http.StatusInternalServerError)
		return
	}
	http.Error(writer, "application request failed", http.StatusBadRequest)
}

func (handler *TenantMemberLifecycleOperationHandler) handleOperationActivateTenantMember(writer http.ResponseWriter, request *http.Request) {
	wire := &accessv1.ActivateTenantMemberRequest{}
	body, err := io.ReadAll(request.Body)
	if err != nil {
		http.Error(writer, "invalid request body", http.StatusBadRequest)
		return
	}
	if len(body) > 0 {
		if err := protojson.Unmarshal(body, wire); err != nil {
			http.Error(writer, "invalid request body", http.StatusBadRequest)
			return
		}
	}
	wire.UserId = request.PathValue("user_id")
	callContext := execution.WithIdempotencyKey(request.Context(), request.Header.Get("Idempotency-Key"))
	output, err := operation.ExecuteTyped(callContext, handler.executor, policy.OperationPlanTenantMemberLifecycleActivateTenantMember(), wire, handler.application.ActivateTenantMember)
	if err != nil {
		writeTenantMemberLifecycleOperationError(writer, err)
		return
	}
	payload, err := protojson.Marshal(output)
	if err != nil {
		http.Error(writer, "response encoding failed", http.StatusInternalServerError)
		return
	}
	writer.Header().Set("Content-Type", "application/json")
	_, _ = writer.Write(payload)
}

func (handler *TenantMemberLifecycleOperationHandler) handleOperationGetTenantMember(writer http.ResponseWriter, request *http.Request) {
	wire := &accessv1.GetTenantMemberRequest{}
	wire.UserId = request.PathValue("user_id")
	callContext := execution.WithIdempotencyKey(request.Context(), request.Header.Get("Idempotency-Key"))
	output, err := operation.ExecuteTyped(callContext, handler.executor, policy.OperationPlanTenantMemberLifecycleGetTenantMember(), wire, handler.application.GetTenantMember)
	if err != nil {
		writeTenantMemberLifecycleOperationError(writer, err)
		return
	}
	payload, err := protojson.Marshal(output)
	if err != nil {
		http.Error(writer, "response encoding failed", http.StatusInternalServerError)
		return
	}
	writer.Header().Set("Content-Type", "application/json")
	_, _ = writer.Write(payload)
}

func (handler *TenantMemberLifecycleOperationHandler) handleOperationInviteTenantMember(writer http.ResponseWriter, request *http.Request) {
	wire := &accessv1.InviteTenantMemberRequest{}
	body, err := io.ReadAll(request.Body)
	if err != nil {
		http.Error(writer, "invalid request body", http.StatusBadRequest)
		return
	}
	if len(body) > 0 {
		if err := protojson.Unmarshal(body, wire); err != nil {
			http.Error(writer, "invalid request body", http.StatusBadRequest)
			return
		}
	}
	callContext := execution.WithIdempotencyKey(request.Context(), request.Header.Get("Idempotency-Key"))
	output, err := operation.ExecuteTyped(callContext, handler.executor, policy.OperationPlanTenantMemberLifecycleInviteTenantMember(), wire, handler.application.InviteTenantMember)
	if err != nil {
		writeTenantMemberLifecycleOperationError(writer, err)
		return
	}
	payload, err := protojson.Marshal(output)
	if err != nil {
		http.Error(writer, "response encoding failed", http.StatusInternalServerError)
		return
	}
	writer.Header().Set("Content-Type", "application/json")
	_, _ = writer.Write(payload)
}

func (handler *TenantMemberLifecycleOperationHandler) handleOperationListTenantMembers(writer http.ResponseWriter, request *http.Request) {
	wire := &accessv1.ListTenantMembersRequest{}
	callContext := execution.WithIdempotencyKey(request.Context(), request.Header.Get("Idempotency-Key"))
	output, err := operation.ExecuteTyped(callContext, handler.executor, policy.OperationPlanTenantMemberLifecycleListTenantMembers(), wire, handler.application.ListTenantMembers)
	if err != nil {
		writeTenantMemberLifecycleOperationError(writer, err)
		return
	}
	payload, err := protojson.Marshal(output)
	if err != nil {
		http.Error(writer, "response encoding failed", http.StatusInternalServerError)
		return
	}
	writer.Header().Set("Content-Type", "application/json")
	_, _ = writer.Write(payload)
}

func (handler *TenantMemberLifecycleOperationHandler) handleOperationRemoveTenantMember(writer http.ResponseWriter, request *http.Request) {
	wire := &accessv1.RemoveTenantMemberRequest{}
	body, err := io.ReadAll(request.Body)
	if err != nil {
		http.Error(writer, "invalid request body", http.StatusBadRequest)
		return
	}
	if len(body) > 0 {
		if err := protojson.Unmarshal(body, wire); err != nil {
			http.Error(writer, "invalid request body", http.StatusBadRequest)
			return
		}
	}
	wire.UserId = request.PathValue("user_id")
	callContext := execution.WithIdempotencyKey(request.Context(), request.Header.Get("Idempotency-Key"))
	output, err := operation.ExecuteTyped(callContext, handler.executor, policy.OperationPlanTenantMemberLifecycleRemoveTenantMember(), wire, handler.application.RemoveTenantMember)
	if err != nil {
		writeTenantMemberLifecycleOperationError(writer, err)
		return
	}
	payload, err := protojson.Marshal(output)
	if err != nil {
		http.Error(writer, "response encoding failed", http.StatusInternalServerError)
		return
	}
	writer.Header().Set("Content-Type", "application/json")
	_, _ = writer.Write(payload)
}

func (handler *TenantMemberLifecycleOperationHandler) handleOperationSuspendTenantMember(writer http.ResponseWriter, request *http.Request) {
	wire := &accessv1.SuspendTenantMemberRequest{}
	body, err := io.ReadAll(request.Body)
	if err != nil {
		http.Error(writer, "invalid request body", http.StatusBadRequest)
		return
	}
	if len(body) > 0 {
		if err := protojson.Unmarshal(body, wire); err != nil {
			http.Error(writer, "invalid request body", http.StatusBadRequest)
			return
		}
	}
	wire.UserId = request.PathValue("user_id")
	callContext := execution.WithIdempotencyKey(request.Context(), request.Header.Get("Idempotency-Key"))
	output, err := operation.ExecuteTyped(callContext, handler.executor, policy.OperationPlanTenantMemberLifecycleSuspendTenantMember(), wire, handler.application.SuspendTenantMember)
	if err != nil {
		writeTenantMemberLifecycleOperationError(writer, err)
		return
	}
	payload, err := protojson.Marshal(output)
	if err != nil {
		http.Error(writer, "response encoding failed", http.StatusInternalServerError)
		return
	}
	writer.Header().Set("Content-Type", "application/json")
	_, _ = writer.Write(payload)
}